requirements.txt
openai
anthropic
google-generativeai
pandas
matplotlib
python-dotenv
termcolor

